main: Avoid panics when walking the AST of body funcs
There were some assumptions on expression types when walking the AST of
the functions body ('hasFundSetCalls' function) based on the tests.

While testing with a real source code the panics happen due to the type
assertions without OK checking, so new assertions have been set for
avoiding the panics and marked with a TODO for future improvement.